+++ spec/playlist_item/status_spec.rb
@@ -10,4 +10,12 @@ describe 'Yt::PlaylistItem’s status methods', :server do
       expect(item.privacy_status).to eq 'public'
     end
   end
+
+  context 'given an item ID for a removed video' do
+    let(:attrs) { {id: $removed_item_id} }
+
+    specify 'return nil with one HTTP call', requests: 1 do
+      expect(item.privacy_status).to be_nil
+    end
+  end
 end